Reconcile many duplicate ways of adding styles into one source, within eddie.geoserver.raster_layers
