+++ src/eddie/geoserver/__init__.py
@@ -20,9 +20,9 @@ Functions and utilities for loading and serving data with geoserver.
 Imports here are accessible directly by `from eddie import geoserver`.
 """
 from .database_layers import create_datastore_layer, create_db_store_if_not_exists, create_main_db_store
-from .geoserver_common import create_workspace_if_not_exists, get_geoserver_url, style_exists
+from .geoserver_common import create_workspace_if_not_exists, get_geoserver_url
 from .terria_catalogs import get_terria_catalog, Workspaces
-from .raster_layers import add_gtiff_to_geoserver, add_style
+from .raster_layers import add_gtiff_to_geoserver, add_style, style_exists
 
 __all__ = [
     "add_gtiff_to_geoserver",
