+++ src/eddie/geoserver/geoserver_common.py
@@ -79,75 +79,3 @@ def create_workspace_if_not_exists(workspace_name: str) -> None:
         # If it does not meet the expected results then raise an error
         # Raise error manually so we can configure the text
         raise requests.HTTPError(response.text, response=response)
-
-
-def style_exists(style_name: str) -> bool:
-    """
-    Check if a GeoServer style definition already exists for a given style_name.
-    The style definition may be empty.
-
-    Parameters
-    ----------
-    style_name : str
-        The name of the style to check for
-
-    Returns
-    -------
-    bool
-        True if the style exists, although it may be empty.
-        False if it does not exist.
-
-    Raises
-    -------
-    HTTPError
-        If geoserver responds with anything but OK or NOT_FOUND, raises it as an exception since it is unexpected.
-    """
-    response = requests.get(
-        f'{get_geoserver_url()}/styles/{style_name}.sld',
-        auth=(EnvVariable.GEOSERVER_ADMIN_NAME, EnvVariable.GEOSERVER_ADMIN_PASSWORD)
-    )
-    if response.status_code == HTTPStatus.OK:
-        return True
-    if response.status_code == HTTPStatus.NOT_FOUND:
-        return False
-    # If it does not meet the expected results then raise an error
-    # Raise error manually, so we can configure the text
-    raise requests.HTTPError(response.text, response=response)
-
-
-def add_style(style_name: str, style_xml: bytes) -> None:
-    """#todo"""
-    log.info(f"Creating style '{style_name}.sld' if it does not exist.")
-    if geoserver.style_exists(style_name):
-        log.debug(f"Style '{style_name}.sld' already exists.")
-    else:
-        # Create the style base
-        create_style_data = f"""
-           <style>
-               <name>{style_name}</name>
-               <filename>{style_name}.sld</filename>
-           </style>
-           """
-        create_style_response = requests.post(
-            f'{geoserver.get_geoserver_url()}/styles',
-            data=create_style_data,
-            headers=_xml_header,
-            auth=(EnvVariable.GEOSERVER_ADMIN_NAME, EnvVariable.GEOSERVER_ADMIN_PASSWORD)
-        )
-        create_style_response.raise_for_status()
-    # PUT the style definition .sld file into the style base
-    sld_response = requests.put(
-        f'{ggeoserver.get_geoserver_url()}/styles/{style_name}',
-        data=style_xml,
-        headers={"Content-type": "application/vnd.ogc.sld+xml"},
-        auth=(EnvVariable.GEOSERVER_ADMIN_NAME, EnvVariable.GEOSERVER_ADMIN_PASSWORD)
-    )
-    sld_response.raise_for_status()
-    log.info(f"Style '{style_name}.sld' created.")
-
-
-def create_viridis_style_if_not_exists() -> None:
-    """Create a GeoServer style for flood rasters using the viridis color scale."""
-    style_name = "viridis_raster"
-    with open('./src/eddie_floodresilience/flood_model/templates/viridis_raster.sld', 'rb') as payload:
-        add_style(style_name, payload)
