+++ src/eddie/geoserver/templates/geotiff_coverage_template.xml
